requirements.txt
asgiref==3.8.1
Django==5.0.6
django-ckeditor==6.7.1
django-filter==24.2
django-js-asset==2.2.0
djangorestframework==3.15.1
Markdown==3.6
pillow==10.3.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
